+++ lib/util/extensions/extensions.dart
@@ -187,6 +187,7 @@ extension ObjectExt<T> on T {
 
 extension FutureSpriteAnimationExt on FutureOr<SpriteAnimation> {
   widget.Widget asWidget({
+    widget.Key? key,
     bool playing = true,
     Anchor anchor = Anchor.topLeft,
     widget.WidgetBuilder? errorBuilder,
@@ -194,26 +195,94 @@ extension FutureSpriteAnimationExt on FutureOr<SpriteAnimation> {
   }) {
     if (this is Future) {
       return widget.FutureBuilder<SpriteAnimation>(
+        key: key,
         future: this as Future<SpriteAnimation>,
-        builder: (context, anim) {
-          if (!anim.hasData) return widget.SizedBox.shrink();
-          return SpriteAnimationWidget(
-            animation: anim.data!,
-            playing: playing,
-            anchor: anchor,
-            errorBuilder: errorBuilder,
-            loadingBuilder: loadingBuilder,
+        builder: (context, data) {
+          if (!data.hasData) return widget.SizedBox.shrink();
+          return widget.Container(
+            constraints: widget.BoxConstraints(
+              minWidth: data.data!.frames.first.sprite.src.width,
+              minHeight: data.data!.frames.first.sprite.src.height,
+            ),
+            child: SpriteAnimationWidget(
+              animation: data.data!,
+              playing: playing,
+              anchor: anchor,
+              errorBuilder: errorBuilder,
+              loadingBuilder: loadingBuilder,
+            ),
           );
         },
       );
     }
 
-    return SpriteAnimationWidget(
-      animation: this as SpriteAnimation ,
-      playing: playing,
-      anchor: anchor,
-      errorBuilder: errorBuilder,
-      loadingBuilder: loadingBuilder,
+    return widget.Container(
+      key: key,
+      constraints: widget.BoxConstraints(
+        minWidth: (this as SpriteAnimation).frames.first.sprite.src.width,
+        minHeight: (this as SpriteAnimation).frames.first.sprite.src.height,
+      ),
+      child: SpriteAnimationWidget(
+        animation: this as SpriteAnimation,
+        playing: playing,
+        anchor: anchor,
+        errorBuilder: errorBuilder,
+        loadingBuilder: loadingBuilder,
+      ),
+    );
+  }
+}
+
+extension FutureSpriteExt on FutureOr<Sprite> {
+  widget.Widget asWidget({
+    widget.Key? key,
+    Anchor anchor = Anchor.topLeft,
+    widget.WidgetBuilder? errorBuilder,
+    widget.WidgetBuilder? loadingBuilder,
+    double angle = 0,
+    Vector2? srcPosition,
+    Vector2? srcSize,
+  }) {
+    if (this is Future) {
+      return widget.FutureBuilder<Sprite>(
+        key: key,
+        future: this as Future<Sprite>,
+        builder: (context, data) {
+          if (!data.hasData) return widget.SizedBox.shrink();
+          return widget.Container(
+            constraints: widget.BoxConstraints(
+              minWidth: data.data!.src.width,
+              minHeight: data.data!.src.height,
+            ),
+            child: SpriteWidget(
+              sprite: data.data!,
+              anchor: anchor,
+              errorBuilder: errorBuilder,
+              loadingBuilder: loadingBuilder,
+              angle: angle,
+              srcPosition: srcPosition,
+              srcSize: srcSize,
+            ),
+          );
+        },
+      );
+    }
+
+    return widget.Container(
+      key: key,
+      constraints: widget.BoxConstraints(
+        minWidth: (this as Sprite).src.width,
+        minHeight: (this as Sprite).src.height,
+      ),
+      child: SpriteWidget(
+        sprite: this as Sprite,
+        anchor: anchor,
+        errorBuilder: errorBuilder,
+        loadingBuilder: loadingBuilder,
+        angle: angle,
+        srcPosition: srcPosition,
+        srcSize: srcSize,
+      ),
     );
   }
 }